# IHM_VRAI_CUSTOM_NET/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebGuitare",
    "version": "1.0"
  },
  "paths": {
...
    "/api/MicroType/GetMicroTypeById": {
      "get": {
        "tags": [
          "MicroType"
        ],
        "operationId": "GetMicroTypeById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_MICROTYPES"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_MICROTYPES"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_MICROTYPES"
                }
              }
            }
          }
        }
      }
    },
    "/api/MicroType/GetMicroTypeByNom": {
      "get": {
        "tags": [
          "MicroType"
        ],
        "operationId": "GetMicroTypeByNom",
        "parameters": [
          {
            "name": "P_Nom",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_MICROTYPES"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_MICROTYPES"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_MICROTYPES"
                }
              }
            }
          }
        }
      }
    },
    "/api/MicroType/AddMicroType": {
      "post": {
        "tags": [
          "MicroType"
        ],
        "operationId": "AddMicroType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_MICROTYPES"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_MICROTYPES"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_MICROTYPES"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/MicroType/UpdateMicroType": {
      "put": {
        "tags": [
          "MicroType"
        ],
        "operationId": "UpdateMicroType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_MICROTYPES"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_MICROTYPES"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_MICROTYPES"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MicroType/DeleteMicroType": {
      "delete": {
        "tags": [
          "MicroType"
        ],
        "operationId": "DeleteMicroType",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Statistique/GetStatistiqueByDateAnnee": {
      "get": {
        "tags": [
          "Statistique"
        ],
        "operationId": "GetStatistiqueByDateAnnee",
        "parameters": [
          {
            "name": "P_Date",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_STATISTIQUE"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_STATISTIQUE"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_STATISTIQUE"
                }
              }
            }
          }
        }
      }
    },
    "/api/Statistique/AddStatistique": {
      "post": {
        "tags": [
          "Statistique"
        ],
        "operationId": "AddStatistique",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_STATISTIQUE"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_STATISTIQUE"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_STATISTIQUE"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Statistique/UpdateStatistique": {
      "put": {
        "tags": [
          "Statistique"
        ],
        "operationId": "UpdateStatistique",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_STATISTIQUE"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_STATISTIQUE"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_STATISTIQUE"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Statistique/DeleteStatistique": {
      "delete": {
        "tags": [
          "Statistique"
        ],
        "operationId": "DeleteStatistique",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/VibratoType/GetVibratoTypeById": {
      "get": {
        "tags": [
          "VibratoType"
        ],
        "operationId": "GetVibratoTypeById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_VIBRATOTYPES"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_VIBRATOTYPES"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_VIBRATOTYPES"
                }
              }
            }
          }
        }
      }
    },
    "/api/VibratoType/GetVibratoTypeByNom": {
      "get": {
        "tags": [
          "VibratoType"
        ],
        "operationId": "GetVibratoTypeByNom",
        "parameters": [
          {
            "name": "P_Nom",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_VIBRATOTYPES"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_VIBRATOTYPES"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_VIBRATOTYPES"
                }
              }
            }
          }
        }
      }
    },
    "/api/VibratoType/AddVibratoType": {
      "post": {
        "tags": [
          "VibratoType"
        ],
        "operationId": "AddVibratoType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_VIBRATOTYPES"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_VIBRATOTYPES"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_VIBRATOTYPES"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/VibratoType/UpdateVibratoType": {
      "put": {
        "tags": [
          "VibratoType"
        ],
        "operationId": "UpdateVibratoType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_VIBRATOTYPES"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_VIBRATOTYPES"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_VIBRATOTYPES"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/VibratoType/DeleteVibratoType": {
      "delete": {
        "tags": [
          "VibratoType"
        ],
        "operationId": "DeleteVibratoType",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "C_MICROTYPES": {
        "type": "object",
        "properties": {
          "idMicro": {
            "type": "integer",
            "format": "int32"
          },
          "nomMicro": {
            "type": "string",
            "nullable": true
          },
          "descriptionMicro": {
            "type": "string",
            "nullable": true
          },
          "dispoMicro": {
            "type": "boolean"
          },
          "prix": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "C_STATISTIQUE": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombreVue": {
            "type": "integer",
            "format": "int32"
          },
          "revenueTotal": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisJanv": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisFevr": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisMars": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisAvri": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisMai": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisJuin": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisJuil": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisAout": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisSept": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisOcto": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisNove": {
            "type": "integer",
            "format": "int32"
          },
          "revenueMoisDece": {
            "type": "integer",
            "format": "int32"
          },
          "revenueAnnee": {
            "type": "integer",
            "format": "int32"
          },
          "dateAnnee": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "C_VIBRATOTYPES": {
        "type": "object",
        "properties": {
          "idVibrato": {
            "type": "integer",
            "format": "int32"
          },
          "nomVibrato": {
            "type": "string",
            "nullable": true
          },
          "descriptionVibrato": {
            "type": "string",
            "nullable": true
          },
          "dispoVibrato": {
            "type": "boolean"
          },
          "prix": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}